rename java filesystem module and package name to "java"
