+++ kivakit-filesystems/java/src/main/java/com/telenav/kivakit/filesystem/java/JavaFile.java
@@ -1,4 +1,22 @@
-package com.telenav.kivakit.filesystem.jfs;
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// © 2011-2021 Telenav, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+package com.telenav.kivakit.filesystem.java;
 
 import com.telenav.kivakit.component.ComponentMixin;
 import com.telenav.kivakit.filesystem.spi.FileService;
@@ -9,7 +27,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Files;
-import java.nio.file.Path;
 
 // @yinyin rename "java-filesystem" to "java" so it's like the others (the project is already under "kivakit-filesystems")
 // @yinyin add module-info.java under src/main/java
@@ -24,12 +41,6 @@ public class JavaFile extends JavaFileSystemObject implements FileService, Compo
     public JavaFile(final String path)
     {
         super(FilePath.parseFilePath(path), false);
-    }
-
-    // @yinyin not used
-    public JavaFile(final Path path)
-    {
-        super(path);
     }
 
     @Override
