+++ kivakit-filesystems/java/src/main/java/com/telenav/kivakit/filesystem/java/JavaFileSystemObject.java
@@ -1,17 +1,38 @@
-package com.telenav.kivakit.filesystem.jfs;
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// © 2011-2021 Telenav, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+package com.telenav.kivakit.filesystem.java;
 
 import com.telenav.kivakit.filesystem.spi.FileSystemObjectService;
 import com.telenav.kivakit.filesystem.spi.FolderService;
 import com.telenav.kivakit.kernel.language.progress.ProgressReporter;
+import com.telenav.kivakit.kernel.language.time.Time;
 import com.telenav.kivakit.resource.CopyMode;
 import com.telenav.kivakit.resource.WritableResource;
 import com.telenav.kivakit.resource.path.FilePath;
 import com.telenav.kivakit.resource.writing.BaseWritableResource;
 
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.attribute.FileTime;
 import java.nio.file.attribute.PosixFilePermission;
 
 import static com.telenav.kivakit.kernel.data.validation.ensure.Ensure.unsupported;
@@ -104,5 +125,20 @@ public class JavaFileSystemObject extends BaseWritableResource implements FileSy
     private static FilePath normalize(final Path path)
     {
         return FilePath.filePath(path);
+    }
+
+    @Override
+    public Time created() {
+        try
+        {
+            FileTime creationTime = (FileTime) Files.getAttribute(this.javaPath, "creationTime");
+            return Time.milliseconds(creationTime.toMillis());
+        }
+        catch (IOException e)
+        {
+            e.printStackTrace();
+        }
+
+        return null;
     }
 }
